+++ src/GameObject.cpp
@@ -5,7 +5,7 @@ GameObject::GameObject()
 
 }
 // initiates update for all children
-void GameObject::update(int dt)
+void GameObject::update(float dt)
 {
     this->onUpdate(dt);
     for(auto a = this->children.begin(); a != this->children.end(); a++){
